+++ lib/pvectl.rb
@@ -89,6 +89,8 @@ require_relative "pvectl/models/syslog_entry"
 require_relative "pvectl/models/journal_entry"
 require_relative "pvectl/models/task_log_line"
 require_relative "pvectl/models/volume"
+require_relative "pvectl/models/dns_config"
+require_relative "pvectl/models/time_config"
 require_relative "pvectl/models/hosts_file"
 
 # Repositories
@@ -106,6 +108,9 @@ require_relative "pvectl/repositories/journal"
 require_relative "pvectl/repositories/task_log"
 require_relative "pvectl/repositories/disk"
 require_relative "pvectl/repositories/volume"
+require_relative "pvectl/repositories/dns"
+require_relative "pvectl/repositories/service"
+require_relative "pvectl/repositories/time_config"
 require_relative "pvectl/repositories/hosts"
 
 # Presenters
@@ -134,6 +139,9 @@ require_relative "pvectl/presenters/journal_entry"
 require_relative "pvectl/presenters/task_log_line"
 require_relative "pvectl/presenters/disk"
 require_relative "pvectl/presenters/volume"
+require_relative "pvectl/presenters/dns_config"
+require_relative "pvectl/presenters/service"
+require_relative "pvectl/presenters/time_config"
 require_relative "pvectl/presenters/hosts_file"
 
 # Selectors
@@ -205,6 +213,9 @@ require_relative "pvectl/services/edit_container"
 # Services - Edit Node
 require_relative "pvectl/services/edit_node"
 
+# Services - Edit DNS
+require_relative "pvectl/services/edit_dns"
+
 # Services - Edit Hosts
 require_relative "pvectl/services/edit_hosts"
 
@@ -244,6 +255,9 @@ require_relative "pvectl/services/pull_config"
 # Services - Push Config
 require_relative "pvectl/services/push_config"
 
+# Services - Service Lifecycle (systemd services on nodes)
+require_relative "pvectl/services/service_lifecycle"
+
 # Commands - Base
 require_relative "pvectl/commands/resource_registry"
 require_relative "pvectl/commands/shared_flags"
@@ -264,6 +278,9 @@ require_relative "pvectl/commands/get/handlers/tasks"
 require_relative "pvectl/commands/get/handlers/templates"
 require_relative "pvectl/commands/get/handlers/disks"
 require_relative "pvectl/commands/get/handlers/volume"
+require_relative "pvectl/commands/get/handlers/dns"
+require_relative "pvectl/commands/get/handlers/services"
+require_relative "pvectl/commands/get/handlers/time"
 require_relative "pvectl/commands/get/handlers/hosts"
 
 # Commands - Describe
@@ -322,6 +339,7 @@ require_relative "pvectl/commands/edit_vm"
 require_relative "pvectl/commands/edit_container"
 require_relative "pvectl/commands/edit_node"
 require_relative "pvectl/commands/edit_volume"
+require_relative "pvectl/commands/edit_dns"
 require_relative "pvectl/commands/edit_hosts"
 
 # Commands - Set
@@ -360,6 +378,9 @@ require_relative "pvectl/commands/pull"
 
 # Commands - Push
 require_relative "pvectl/commands/push"
+
+# Commands - Service (systemd lifecycle on nodes)
+require_relative "pvectl/commands/service"
 
 # Wizards
 require_relative "pvectl/wizards/create_vm"