+++ lib/pvectl/commands/edit_vm.rb
@@ -41,6 +41,9 @@ module Pvectl
             Edit a node configuration:
               $ pvectl edit node pve1
 
+            Edit node DNS resolver settings:
+              $ pvectl edit dns pve1
+
             Edit volume properties:
               $ pvectl edit volume vm 100 scsi0
 
@@ -57,7 +60,11 @@ module Pvectl
             In --dry-run mode, shows the diff between current and edited
             configuration without applying changes to Proxmox.
 
-            Supported resource types: vm, container (ct), node, volume, hosts.
+            Supported resource types: vm, container (ct), node, volume, dns, hosts.
+
+            For DNS, the node identifier is given as a positional argument
+            (pvectl edit dns NODE). DNS settings include search domain and
+            up to three nameservers (dns1, dns2, dns3).
 
             For volumes, syntax is: pvectl edit volume <vm|container> <id> <disk>
 
@@ -95,11 +102,13 @@ module Pvectl
               Commands::EditNode.execute(resource_ids, options, global_options)
             when "volume"
               Commands::EditVolume.execute(resource_ids, options, global_options)
+            when "dns"
+              Commands::EditDns.execute(resource_ids, options, global_options)
             when "hosts"
               Commands::EditHosts.execute(resource_ids, options, global_options)
             else
               $stderr.puts "Error: Unknown resource type: #{resource_type}"
-              $stderr.puts "Valid types: vm, container, node, volume, hosts"
+              $stderr.puts "Valid types: vm, container, node, volume, dns, hosts"
               ExitCodes::USAGE_ERROR
             end